feat: isolate learn mode + autosave + deep analysis logic
