+++ src/contexts/index.ts
@@ -8,5 +8,5 @@ export * from './StockfishEngineContext'
 export * from './TourContext'
 export * from './TrainingControllerContext'
 export * from './TreeControllerContext'
-export * from './TuringTreeControllerContext'
+export * from './TuringControllerContext'
 export * from './WindowSizeContext'
