+++ src/hooks/useAnalysisController/index.ts
@@ -3,3 +3,6 @@ export * from './useEngineAnalysis'
 export * from './useAnalysisController'
 export * from './useMoveRecommendations'
 export * from './useBoardDescription'
+export * from './useLearnFromMistakes'
+export * from './useDeepAnalysis'
+export * from './useAutoSave'
